+++ http4k-serverless-alibaba/integration-test/test-function/src/main/kotlin/org/http4k/serverless/alibaba/TestFunction.kt
@@ -1,8 +1,0 @@
-@file:Suppress("unused")
-
-package org.http4k.serverless.alibaba
-
-import org.http4k.client.ServerForClientContract
-import org.http4k.serverless.AlibabaFunction
-
-class TestFunction : AlibabaFunction(ServerForClientContract)