+++ http4k-serverless-alibaba/integration-test/test-function/src/test/kotlin/org/http4k/serverless/alibaba/Placeholder.kt
@@ -1,3 +1,0 @@
-package org.http4k.serverless.alibaba
-
-object Placeholder